requirements.txt
Flask==2.0.1
gunicorn==20.1.0
fastapi==0.68.0
pydantic==1.8.2
langchain-ollama==0.1.0
langchain-core==0.1.0
flask-wtf==0.15.1
python-dotenv==0.19.0
tabulate==0.8.9
